# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "summary": "Refresh access token",
        "description": "Refresh the access token using the refresh token from cookies. Sets new authentication cookies.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "401": {
            "description": "Unauthorized - invalid refresh token"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Clinic Management API",
    "description": "AI-Powered Pet Clinic Management Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}